# postman/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "API Raizes do Nordeste",
        "description": "Back-end da rede de lanchonetes Raizes do Nordeste.\n\nAtende varias unidades da franquia e varios canais de venda\n(APP, TOTEM, BALCAO, PICKUP e WEB).\n\nComo testar:\n1. Chame POST /auth/login com um usuario do seed.\n2. Clique em Authorize aqui em cima e cole o accessToken.\n3. As rotas protegidas passam a funcionar.\n\nUsuarios do seed (senha Senha@123):\n- admin@raizes.com.br (ADMIN)\n- gerente.recife@raizes.com.br (GERENTE)\n- atendente.recife@raizes.com.br (ATENDENTE)\n- cozinha.recife@raizes.com.br (COZINHA)\n- cliente@exemplo.com (CLIENTE)",
        "contact": {
            "name": "Gean Brandao"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Ambiente local"
        },
        {
            "url": "/",
            "description": "Mesma origem que serviu esta pagina"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
    "paths": {
...
        "/unidades/{unidadeId}/estoque/movimentacoes": {
            "post": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Movimentar estoque",
                "description": "Perfis ATENDENTE, GERENTE e ADMIN.\n\nENTRADA soma ao saldo, SAIDA subtrai e AJUSTE define o saldo absoluto (caso da contagem de inventario).\n\nSe o produto ainda n\u00e3o tiver saldo nesta unidade, a linha e criada na primeira entrada.",
                "operationId": "movimentar",
                "parameters": [
                    {
                        "name": "unidadeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovimentacaoEstoqueRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Movimenta\u00e7\u00e3o registrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MovimentacaoEstoqueResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o, ou operador de outra unidade",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade ou produto n\u00e3o encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Saida maior que o saldo disponivel",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos invalidos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/unidades/{unidadeId}/estoque": {
            "get": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Consultar saldo da unidade",
                "description": "Perfis ATENDENTE, COZINHA, GERENTE e ADMIN. Operador so enxerga a propria unidade; admin enxerga qualquer uma.",
                "operationId": "listarSaldos",
                "parameters": [
                    {
                        "name": "unidadeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saldos da unidade",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginaResponseEstoqueResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o, ou operador de outra unidade",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade n\u00e3o encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/unidades/{unidadeId}/estoque/{produtoId}/movimentacoes": {
            "get": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Hist\u00f3rico de movimenta\u00e7\u00f5es do produto",
                "description": "Perfis GERENTE e ADMIN. Cada linha guarda o saldo que ficou depois dela, ent\u00e3o o hist\u00f3rico e conferivel sem recalcular nada.",
                "operationId": "listarMovimentacoes",
                "parameters": [
                    {
                        "name": "unidadeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "produtoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagina, come\u00e7ando em 1",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hist\u00f3rico do produto",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginaResponseMovimentacaoEstoqueResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o, ou operador de outra unidade",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto sem estoque nesta unidade",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/auditoria": {
            "get": {
                "tags": [
                    "Auditoria"
                ],
                "summary": "Consultar trilha de auditoria",
                "description": "Exclusivo do perfil ADMIN.\n\nA\u00e7\u00f5es rastreadas: cria\u00e7\u00e3o, cancelamento e mudan\u00e7a de status de pedido, movimenta\u00e7\u00e3o de estoque, solicita\u00e7\u00e3o de pagamento, ac\u00famulo e resgate de pontos, registro e revoga\u00e7\u00e3o de consentimento.",
                "operationId": "consultar",
                "parameters": [
                    {
                        "name": "usuarioId",
                        "in": "query",
                        "description": "Filtra por quem fez a a\u00e7\u00e3o",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "entidade",
                        "in": "query",
                        "description": "PEDIDO, ESTOQUE, PAGAMENTO, FIDELIDADE ou CONSENTIMENTO",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registros da trilha",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginaResponseLogAuditoriaResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorDetail": {
                "type": "object",
                "description": "Detalhe de um campo invalido",
                "properties": {
                    "field": {
                        "type": "string",
                        "example": "itens[0].quantidade"
                    },
                    "issue": {
                        "type": "string",
                        "example": "deve ser maior que zero"
                    }
                },
                "required": [
                    "field",
                    "issue"
                ]
            },
            "ErrorResponse": {
                "type": "object",
                "description": "Formato padr\u00e3o de erro da API",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "ESTOQUE_INSUFICIENTE"
                    },
                    "message": {
                        "type": "string",
                        "example": "N\u00e3o ha quantidade suficiente para um ou mais itens."
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorDetail"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-05T12:00:00Z"
                    },
                    "path": {
                        "type": "string",
                        "example": "/pedidos"
                    },
                    "requestId": {
                        "type": "string",
                        "example": "3f1a2b4c-5d6e-7f80-91a2-b3c4d5e6f708"
                    }
                },
                "required": [
                    "details",
                    "error",
                    "message",
                    "timestamp"
                ]
            },
...
            "MovimentacaoEstoqueRequest": {
                "type": "object",
                "description": "Movimenta\u00e7\u00e3o de estoque",
                "properties": {
                    "produtoId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "tipo": {
                        "type": "string",
                        "description": "ENTRADA soma, SAIDA subtrai, AJUSTE define o saldo absoluto",
                        "enum": [
                            "ENTRADA",
                            "SAIDA",
                            "AJUSTE"
                        ],
                        "example": "ENTRADA"
                    },
                    "quantidade": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Em ENTRADA e SAIDA, quanto movimentar (maior que zero). Em AJUSTE, o saldo que passa a valer, e ai zero e aceito: a contagem de inventario pode dar zero.",
                        "example": 20,
                        "minimum": 0
                    },
                    "motivo": {
                        "type": "string",
                        "example": "Recebimento do fornecedor",
                        "maxLength": 255,
                        "minLength": 0
                    }
                },
                "required": [
                    "produtoId",
                    "quantidade",
                    "tipo"
                ]
            },
            "MovimentacaoEstoqueResponse": {
                "type": "object",
                "description": "Movimenta\u00e7\u00e3o registrada",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "produtoId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "tipo": {
                        "type": "string",
                        "enum": [
                            "ENTRADA",
                            "SAIDA",
                            "AJUSTE"
                        ]
                    },
                    "quantidade": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "saldoApos": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Saldo que ficou depois desta movimenta\u00e7\u00e3o",
                        "example": 70
                    },
                    "motivo": {
                        "type": "string"
                    },
                    "pedidoId": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Preenchido quando a movimenta\u00e7\u00e3o veio de um pedido"
                    },
                    "criadoEm": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "criadoEm",
                    "id",
                    "nome",
                    "produtoId",
                    "quantidade",
                    "saldoApos",
                    "tipo"
                ]
            },
...
            "EstoqueResponse": {
                "type": "object",
                "description": "Saldo de estoque de um produto na unidade",
                "properties": {
                    "produtoId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "categoria": {
                        "type": "string"
                    },
                    "saldoAtual": {
                        "type": "integer",
                        "format": "int32",
                        "example": 50
                    },
                    "saldoMinimo": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Abaixo disso a loja deveria repor",
                        "example": 10
                    },
                    "abaixoDoMinimo": {
                        "type": "boolean",
                        "description": "true quando o saldo esta no minimo ou abaixo dele"
                    }
                },
                "required": [
                    "abaixoDoMinimo",
                    "categoria",
                    "nome",
                    "produtoId",
                    "saldoAtual",
                    "saldoMinimo"
                ]
            },
            "PaginaResponseEstoqueResponse": {
                "type": "object",
                "description": "Listagem paginada",
                "properties": {
                    "conteudo": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EstoqueResponse"
                        }
                    },
                    "pagina": {
                        "type": "integer",
                        "format": "int32",
                        "example": 1
                    },
                    "limite": {
                        "type": "integer",
                        "format": "int32",
                        "example": 10
                    },
                    "totalItens": {
                        "type": "integer",
                        "format": "int64",
                        "example": 42
                    },
                    "totalPaginas": {
                        "type": "integer",
                        "format": "int32",
                        "example": 5
                    },
                    "primeira": {
                        "type": "boolean"
                    },
                    "ultima": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "conteudo",
                    "limite",
                    "pagina",
                    "primeira",
                    "totalItens",
                    "totalPaginas",
                    "ultima"
                ]
            },
            "PaginaResponseMovimentacaoEstoqueResponse": {
                "type": "object",
                "description": "Listagem paginada",
                "properties": {
                    "conteudo": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MovimentacaoEstoqueResponse"
                        }
                    },
                    "pagina": {
                        "type": "integer",
                        "format": "int32",
                        "example": 1
                    },
                    "limite": {
                        "type": "integer",
                        "format": "int32",
                        "example": 10
                    },
                    "totalItens": {
                        "type": "integer",
                        "format": "int64",
                        "example": 42
                    },
                    "totalPaginas": {
                        "type": "integer",
                        "format": "int32",
                        "example": 5
                    },
                    "primeira": {
                        "type": "boolean"
                    },
                    "ultima": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "conteudo",
                    "limite",
                    "pagina",
                    "primeira",
                    "totalItens",
                    "totalPaginas",
                    "ultima"
                ]
            },
...
            "LogAuditoriaResponse": {
                "type": "object",
                "description": "Registro de auditoria",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "usuarioId": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Quem fez. Nulo quando a a\u00e7\u00e3o partiu do sistema."
                    },
                    "acao": {
                        "type": "string",
                        "example": "PEDIDO_CANCELADO"
                    },
                    "entidade": {
                        "type": "string",
                        "example": "PEDIDO"
                    },
                    "entidadeId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "dadosAnteriores": {
                        "type": "string",
                        "description": "Estado anterior em JSON. Nulo em cria\u00e7\u00e3o."
                    },
                    "dadosNovos": {
                        "type": "string",
                        "description": "Estado novo em JSON. Nulo em exclus\u00e3o."
                    },
                    "ip": {
                        "type": "string"
                    },
                    "criadoEm": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "acao",
                    "criadoEm",
                    "entidade",
                    "id"
                ]
            },
            "PaginaResponseLogAuditoriaResponse": {
                "type": "object",
                "description": "Listagem paginada",
                "properties": {
                    "conteudo": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LogAuditoriaResponse"
                        }
                    },
                    "pagina": {
                        "type": "integer",
                        "format": "int32",
                        "example": 1
                    },
                    "limite": {
                        "type": "integer",
                        "format": "int32",
                        "example": 10
                    },
                    "totalItens": {
                        "type": "integer",
                        "format": "int64",
                        "example": 42
                    },
                    "totalPaginas": {
                        "type": "integer",
                        "format": "int32",
                        "example": 5
                    },
                    "primeira": {
                        "type": "boolean"
                    },
                    "ultima": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "conteudo",
                    "limite",
                    "pagina",
                    "primeira",
                    "totalItens",
                    "totalPaginas",
                    "ultima"
                ]
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Cole aqui o accessToken devolvido por POST /auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}